Speeduino-February2020 Easy install
Everything Fuel Injection.msi is Speeduino-February2020 Easy install for Speeduino firmware for all my current boards with ATMEGA2560. once your board is plugged in, take note of the comport, when you run this software and its installed it will ask you to select a com port from the drop down window, select the port of your ECU and it will automatically upload to your device. No support for Teensy in this version. Windows only.